CDPS-1540 Additional test splitting and tidying
