+++ integration_tests/e2e/personal/editDistinguishingMarks/viewAllImagesForDistinguishingMark.cy.ts
@@ -1,7 +1,7 @@
-import { Role } from '../../../../../server/data/enums/role'
-import Page from '../../../../pages/page'
-import { distinguishingMarkMultiplePhotosMock } from '../../../../../server/data/localMockData/distinguishingMarksMock'
-import ViewAllImagesForDistinguishingMark from '../../../../pages/editPages/distinguishingMarks/viewAllImagesForDistinguishingMark'
+import { Role } from '../../../../server/data/enums/role'
+import Page from '../../../pages/page'
+import { distinguishingMarkMultiplePhotosMock } from '../../../../server/data/localMockData/distinguishingMarksMock'
+import ViewAllImagesForDistinguishingMark from '../../../pages/editPages/distinguishingMarks/viewAllImagesForDistinguishingMark'
 
 const prisonerNumber = 'G6123VU'
 const bookingId = 1102484
