+++ package.json
@@ -22,8 +22,9 @@
     "security_audit": "npx audit-ci --config audit-ci.json",
     "int-test": "cypress run",
     "int-test-ui": "cypress open --browser chrome",
-    "int-test-ci-1": "cypress run --spec \"**/personal/edit/**/*.cy.ts\"",
-    "int-test-ci-2": "cypress run --spec \"**/e2e/*.cy.ts,**/alerts/**/*.cy.ts,**/personal/*.cy.ts,**/photo/**/*,**/workAndSkillsPage/**/*.cy.ts\"",
+    "integration-test-1": "cypress run --spec \"**/personal/edit/**/*.cy.ts\"",
+    "integration-test-2": "cypress run --spec \"**/personal/editDistinguishingMarks/**/*.cy.ts\"",
+    "integration-test-3": "cypress run --spec \"**/e2e/*.cy.ts,**/alerts/**/*.cy.ts,**/personal/*.cy.ts,**/photo/**/*,**/workAndSkillsPage/**/*.cy.ts\"",
     "clean": "rm -rf dist test_results",
     "rebuild": "npm run clean && npm i && npm run build",
     "postinstall": "npx patch-package",
